# Conflicts:
#	doc/user_guide/boosting/parameters.md
#	doc/user_guide/seco/parameters.md

+++ cpp/subprojects/seco/src/mlrl/seco/heuristics/heuristic_common.hpp
@@ -8,26 +8,26 @@
 
 namespace seco {
 
-    static inline constexpr float64 precision(float64 cin, float64 cip, float64 crn, float64 crp) {
-        float64 numCoveredCorrect = cin + crp;
-        float64 numCovered = numCoveredCorrect + cip + crn;
+    static inline constexpr float32 precision(float32 cin, float32 cip, float32 crn, float32 crp) {
+        float32 numCoveredCorrect = cin + crp;
+        float32 numCovered = numCoveredCorrect + cip + crn;
         return util::divideOrZero(numCoveredCorrect, numCovered);
     }
 
-    static inline constexpr float64 recall(float64 cin, float64 crp, float64 uin, float64 urp) {
-        float64 numCoveredEqual = cin + crp;
-        float64 numEqual = numCoveredEqual + uin + urp;
+    static inline constexpr float32 recall(float32 cin, float32 crp, float32 uin, float32 urp) {
+        float32 numCoveredEqual = cin + crp;
+        float32 numEqual = numCoveredEqual + uin + urp;
         return util::divideOrZero(numCoveredEqual, numEqual);
     }
 
-    static inline constexpr float64 wra(float64 cin, float64 cip, float64 crn, float64 crp, float64 uin, float64 uip,
-                                        float64 urn, float64 urp) {
-        float64 numCoveredEqual = cin + crp;
-        float64 numUncoveredEqual = uin + urp;
-        float64 numEqual = numUncoveredEqual + numCoveredEqual;
-        float64 numCovered = numCoveredEqual + cip + crn;
-        float64 numUncovered = numUncoveredEqual + uip + urn;
-        float64 numTotal = numCovered + numUncovered;
+    static inline constexpr float32 wra(float32 cin, float32 cip, float32 crn, float32 crp, float32 uin, float32 uip,
+                                        float32 urn, float32 urp) {
+        float32 numCoveredEqual = cin + crp;
+        float32 numUncoveredEqual = uin + urp;
+        float32 numEqual = numUncoveredEqual + numCoveredEqual;
+        float32 numCovered = numCoveredEqual + cip + crn;
+        float32 numUncovered = numUncoveredEqual + uip + urn;
+        float32 numTotal = numCovered + numUncovered;
 
         if (numCovered > 0 && numTotal > 0) {
             return (numCovered / numTotal) * ((numCoveredEqual / numCovered) - (numEqual / numTotal));
